+++ tests/test_bandpass.py
@@ -276,6 +276,7 @@ def test_ne():
 
 @timer
 def test_thin():
+    """Test that bandpass thinning works with the requested accuracy."""
     s = galsim.SED('1', wave_type='nm', flux_type='fphotons')
     bp = galsim.Bandpass(os.path.join(datapath, 'LSST_r.dat'), 'nm')
     flux = s.calculateFlux(bp)
@@ -300,6 +301,50 @@ def test_thin():
         print("realized error = ",(flux-thin_flux)/flux)
         assert np.abs(thin_err) < err, "Thinned bandpass failed accuracy goal, w/ range shrinkage."
 
+@timer
+def test_zp():
+    """Check that the zero points are maintained in an appropriate way when thinning, truncating."""
+    # Make a bandpass and set an AB zeropoint.
+    bp = galsim.Bandpass(os.path.join(datapath, 'LSST_r.dat'), 'nm')
+    bp = bp.withZeropoint(zeropoint='AB', effective_diameter=6.4, exptime=15)
+    # Confirm that if we use the default thinning kwargs, then the zeropoint for the thinned
+    # bandpass is the same (exactly) as the original.
+    bp_th = bp.thin()
+    np.testing.assert_equal(bp.zeropoint, bp_th.zeropoint,
+                            "Zeropoint not preserved after thinning with defaults")
+    bp_tr = bp.truncate(relative_throughput=1.e-4)
+    np.testing.assert_equal(bp.zeropoint, bp_tr.zeropoint,
+                            "Zeropoint not preserved after truncating with defaults")
+
+    # Confirm that if we explicit set the kwarg to clear the zeropoint when thinning or truncating,
+    # or if we truncate using blue_limit or red_limit, then the new bandpass has no zeropoint
+    bp_th = bp.thin(preserve_zp = False)
+    assert bp_th.zeropoint is None, \
+        "Zeropoint erroneously preserved after thinning with preserve_zp=False"
+    bp_tr = bp.truncate(preserve_zp = False)
+    assert bp_tr.zeropoint is None, \
+        "Zeropoint erroneously preserved after truncating with preserve_zp=False"
+    bp_tr = bp.truncate(red_limit = 600.)
+    assert bp_tr.zeropoint is None, \
+        "Zeropoint erroneously preserved after truncating with explicit red_limit"
+    bp_tr = bp.truncate(blue_limit = 500.)
+    assert bp_tr.zeropoint is None, \
+        "Zeropoint erroneously preserved after truncating with explicit blue_limit"
+
+@timer
+def test_truncate_inputs():
+    """Test that bandpass truncation respects certain sanity constraints on the inputs."""
+    try:
+        # Don't allow truncation via two different methods.
+        bp = galsim.Bandpass(os.path.join(datapath, 'LSST_r.dat'), 'nm')
+        np.testing.assert_raises(ValueError, bp.truncate, relative_throughput=1.e-4, blue_limit=500.)
+
+        # If blue_limit or red_limit is supplied, don't allow values that are outside the original
+        # wavelength range.
+        np.testing.assert_raises(ValueError, bp.truncate, blue_limit=0.9*bp.blue_limit)
+        np.testing.assert_raises(ValueError, bp.truncate, red_limit=1.1*bp.red_limit)
+    except ImportError:
+        print 'The assert_raises tests require nose'
 
 if __name__ == "__main__":
     test_Bandpass_basic()
@@ -308,3 +353,5 @@ if __name__ == "__main__":
     test_Bandpass_wave_type()
     test_ne()
     test_thin()
+    test_zp()
+    test_truncate_inputs()