+++ CHANGELOG.md
@@ -66,7 +66,6 @@ Bug Fixes
   adding or convolving a single element that could lead to erroneous str and
   reprs for the resulting object if it was then transformed. (#763)
 
-
 Deprecated Features
 -------------------
 
@@ -123,6 +122,8 @@ New Features
 - Added the FourierSqrt operator to compute the Fourier-space square root of a
   profile.  This is useful in implementing optimal coaddition algorithms; see
   make_coadd.py in the examples directory (#748).
+- Made Bandpass.thin() and Bandpass.truncate() preserve the zeropoint by 
+  default. (#711)
 - Added version information to the compiled C++ library. (#750)